add request info on the custom close form
